requirements.txt
pyaudio
wave
python-dotenv
openai
groq
webrtcvad
pydub
simpleaudio